Revert "Wumpus Collision Changed"
