+++ Assets/Materials/GoldTiles.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6e87f441c65c18f40b60136856f8e8f1
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
