+++ Assets/Materials/RustedMetal.mat.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: a4b5186ef0bcb7f4b898d096622b180f
-PrefabImporter:
+guid: 13f7a11a3e9a18044a6276d957f83886
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 2100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
