+++ Assets/textures.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: bbb769f343af70f488b62238249eb1b8
-PrefabImporter:
+guid: cfdcce6ccbd098f4aa027579ff37eada
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
